FIX/FEATURE: Add mock and types directories. Cleaning old typings etc..
