+++ frontend/src/services/user/users.ts
@@ -1,19 +1,8 @@
 import axios from 'axios';
-import {CustomImage} from '../image/images';
 import {getAccessToken} from '../utils';
+import {CustomImage, CustomUser} from '../../types/types';
 
-const API_URL = 'http://localhost:8000/api/users/';
-
-export type User = {
-  id: number;
-  username: string;
-  email: string;
-  bio?: string;
-  num_followers?: number;
-  num_following?: number;
-  profile_picture_url?: string;
-  num_posts?: number;
-};
+const API_URL = 'http://localhost:8000/api/CustomUsers/';
 
 const instance = axios.create({
   baseURL: API_URL,
@@ -28,50 +17,52 @@ instance.interceptors.request.use(async config => {
   return config;
 });
 
-export const getUsers = async (): Promise<User[]> => {
-  const response = await instance.get<User[]>('/');
+export const getCustomCustomUsers = async (): Promise<CustomUser[]> => {
+  const response = await instance.get<CustomUser[]>('/');
   return response.data;
 };
 
-export const getUserById = async (id: number): Promise<User> => {
-  const response = await instance.get<User>(`${id}/`);
+export const getCustomUserById = async (id: number): Promise<CustomUser> => {
+  const response = await instance.get<CustomUser>(`${id}/`);
   return response.data;
 };
 
-export const updateUser = async (
+export const updateCustomUser = async (
   id: number,
-  userData: Partial<User>,
-): Promise<User> => {
-  const response = await instance.put<User>(`${id}/`, userData);
+  CustomUserData: Partial<CustomUser>,
+): Promise<CustomUser> => {
+  const response = await instance.put<CustomUser>(`${id}/`, CustomUserData);
   return response.data;
 };
 
-export const deleteUser = async (id: number): Promise<void> => {
+export const deleteCustomUser = async (id: number): Promise<void> => {
   await instance.delete(`${id}/`);
 };
 
-export const getImagesByUser = async (id: number): Promise<CustomImage[]> => {
+export const getImagesByCustomUser = async (
+  id: number,
+): Promise<CustomImage[]> => {
   const response = await instance.get<CustomImage[]>(`${id}/images/`);
   return response.data;
 };
 /* not implemented in backend|
 ------------------------------------------------------------------------------
-Todo: table for storing followers relations: id: PK, user_id1: FK, user_id2: FK;
+Todo: table for storing followers relations: id: PK, CustomUser_id1: FK, CustomUser_id2: FK;
 
 /api/
-    ├── users/
-    │   ├── {id}/                  # Get, update, or delete a specific user
-    │   ├── {id}/follow/           # Follow a user
-    │   ├── {id}/unfollow/         # Unfollow a user
-    │   ├── {id}/followers/        # Get list of followers for a user
-    │   └── {id}/following/        # Get list of users a user is following
+    ├── CustomUsers/
+    │   ├── {id}/                  # Get, update, or delete a specific CustomUser
+    │   ├── {id}/follow/           # Follow a CustomUser
+    │   ├── {id}/unfollow/         # Unfollow a CustomUser
+    │   ├── {id}/followers/        # Get list of followers for a CustomUser
+    │   └── {id}/following/        # Get list of CustomUsers a CustomUser is following
 -----------------------------------------------------------------------------
-export const getUserFollowers = async (id: number): Promise<CustomImage[]> => {
+export const getCustomUserFollowers = async (id: number): Promise<CustomImage[]> => {
   const response = await instance.get<CustomImage[]>(`${id}/`);
   return response.data;
 };
 
-export const getUserFollowing = async (id: number): Promise<CustomImage[]> => {
+export const getCustomUserFollowing = async (id: number): Promise<CustomImage[]> => {
   const response = await instance.get<CustomImage[]>(`${id}/`);
   return response.data;
 };
